requirements.txt
numpy==1.14.3
PyYAML==3.12
scikit-image==0.13.1
scikit-learn==0.20.0
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0


